ticketvote: Add CmdDetails and CmdResults.
