+++ politeiad/backendv2/tstorebe/plugins/ticketvote/cmds.go
@@ -26,185 +26,6 @@ import (
 	"github.com/decred/politeia/util"
 )
 
-const (
-	pluginID = ticketvote.PluginID
-
-	// Blob entry data descriptors
-	dataDescriptorAuthDetails     = pluginID + "-auth-v1"
-	dataDescriptorVoteDetails     = pluginID + "-vote-v1"
-	dataDescriptorCastVoteDetails = pluginID + "-castvote-v1"
-	dataDescriptorVoteCollider    = pluginID + "-vcollider-v1"
-	dataDescriptorStartRunoff     = pluginID + "-startrunoff-v1"
-)
-
-// cmdAuthorize authorizes a ticket vote or revokes a previous authorization.
-func (p *plugin) cmdAuthorize(token []byte, payload string) (string, error) {
-	// Decode payload
-	var a ticketvote.Authorize
-	err := json.Unmarshal([]byte(payload), &a)
-	if err != nil {
-		return "", err
-	}
-
-	// Verify token
-	err = tokenMatches(token, a.Token)
-	if err != nil {
-		return "", err
-	}
-
-	// Verify signature
-	version := strconv.FormatUint(uint64(a.Version), 10)
-	msg := a.Token + version + string(a.Action)
-	err = util.VerifySignature(a.Signature, a.PublicKey, msg)
-	if err != nil {
-		return "", convertSignatureError(err)
-	}
-
-	// Verify action
-	switch a.Action {
-	case ticketvote.AuthActionAuthorize:
-		// This is allowed
-	case ticketvote.AuthActionRevoke:
-		// This is allowed
-	default:
-		return "", backend.PluginError{
-			PluginID:  ticketvote.PluginID,
-			ErrorCode: uint32(ticketvote.ErrorCodeAuthorizationInvalid),
-			ErrorContext: fmt.Sprintf("%v not a valid action",
-				a.Action),
-		}
-	}
-
-	// Verify record status and version
-	r, err := p.tstore.RecordPartial(token, 0, nil, true)
-	if err != nil {
-		return "", errors.Errorf("RecordPartial: %v", err)
-	}
-	if r.RecordMetadata.Status != backend.StatusPublic {
-		return "", backend.PluginError{
-			PluginID:     ticketvote.PluginID,
-			ErrorCode:    uint32(ticketvote.ErrorCodeRecordStatusInvalid),
-			ErrorContext: "record is not public",
-		}
-	}
-	if a.Version != r.RecordMetadata.Version {
-		return "", backend.PluginError{
-			PluginID:  ticketvote.PluginID,
-			ErrorCode: uint32(ticketvote.ErrorCodeRecordVersionInvalid),
-			ErrorContext: fmt.Sprintf("version is not latest: "+
-				"got %v, want %v", a.Version,
-				r.RecordMetadata.Version),
-		}
-	}
-
-	// Get any previous authorizations to verify that the new action
-	// is allowed based on the previous action.
-	auths, err := p.auths(token)
-	if err != nil {
-		return "", err
-	}
-	var prevAction ticketvote.AuthActionT
-	if len(auths) > 0 {
-		prevAction = ticketvote.AuthActionT(auths[len(auths)-1].Action)
-	}
-	switch {
-	case len(auths) == 0:
-		// No previous actions. New action must be an authorize.
-		if a.Action != ticketvote.AuthActionAuthorize {
-			return "", backend.PluginError{
-				PluginID:  ticketvote.PluginID,
-				ErrorCode: uint32(ticketvote.ErrorCodeAuthorizationInvalid),
-				ErrorContext: "no prev action; action must " +
-					"be authorize",
-			}
-		}
-	case prevAction == ticketvote.AuthActionAuthorize &&
-		a.Action != ticketvote.AuthActionRevoke:
-		// Previous action was a authorize. This action must be revoke.
-		return "", backend.PluginError{
-			PluginID:     ticketvote.PluginID,
-			ErrorCode:    uint32(ticketvote.ErrorCodeAuthorizationInvalid),
-			ErrorContext: "prev action was authorize",
-		}
-	case prevAction == ticketvote.AuthActionRevoke &&
-		a.Action != ticketvote.AuthActionAuthorize:
-		// Previous action was a revoke. This action must be authorize.
-		return "", backend.PluginError{
-			PluginID:     ticketvote.PluginID,
-			ErrorCode:    uint32(ticketvote.ErrorCodeAuthorizationInvalid),
-			ErrorContext: "prev action was revoke",
-		}
-	}
-
-	// Prepare authorize vote
-	receipt := p.identity.SignMessage([]byte(a.Signature))
-	auth := ticketvote.AuthDetails{
-		Token:     a.Token,
-		Version:   a.Version,
-		Action:    string(a.Action),
-		PublicKey: a.PublicKey,
-		Signature: a.Signature,
-		Timestamp: time.Now().Unix(),
-		Receipt:   hex.EncodeToString(receipt[:]),
-	}
-
-	// Save authorize vote
-	err = p.authSave(token, auth)
-	if err != nil {
-		return "", err
-	}
-
-	// Update inventory
-	var status ticketvote.VoteStatusT
-	switch a.Action {
-	case ticketvote.AuthActionAuthorize:
-		status = ticketvote.VoteStatusAuthorized
-	case ticketvote.AuthActionRevoke:
-		status = ticketvote.VoteStatusUnauthorized
-	default:
-		// Action has already been validated. This should not happen.
-		return "", errors.Errorf("invalid action %v", a.Action)
-	}
-	p.inventoryUpdate(a.Token, status)
-
-	// Prepare reply
-	ar := ticketvote.AuthorizeReply{
-		Timestamp: auth.Timestamp,
-		Receipt:   auth.Receipt,
-	}
-	reply, err := json.Marshal(ar)
-	if err != nil {
-		return "", err
-	}
-
-	return string(reply), nil
-}
-
-// voteCollider is used to prevent duplicate votes at the tlog level. The
-// backend saves a digest of the data to the trillian log (tlog). Tlog does not
-// allow leaves with duplicate values, so once a vote colider is saved to the
-// backend for a ticket it should be impossible for another vote collider to be
-// saved to the backend that is voting with the same ticket on the same record,
-// regardless of what the vote bits are. The vote collider and the full cast
-// vote are saved to the backend at the same time. A cast vote is not
-// considered valid unless a corresponding vote collider is present.
-type voteCollider struct {
-	Token  string `json:"token"`  // Record token
-	Ticket string `json:"ticket"` // Ticket hash
-}
-
-// voteColliderSave saves a voteCollider to the backend.
-func (p *plugin) voteColliderSave(token []byte, vc voteCollider) error {
-	// Prepare blob
-	be, err := convertBlobEntryFromVoteCollider(vc)
-	if err != nil {
-		return err
-	}
-
-	// Save blob
-	return tstore.BlobSave(token, *be)
-}
-
 // ballotResults is used to aggregate data for votes that are cast
 // concurrently.
 type ballotResults struct {
@@ -754,33 +575,6 @@ func (p *plugin) cmdCastBallot(token []byte, payload string) (string, error) {
 	return string(reply), nil
 }
 
-// cmdDetails returns the vote details for a record.
-func (p *plugin) cmdDetails(token []byte) (string, error) {
-	// Get vote authorizations
-	auths, err := p.auths(token)
-	if err != nil {
-		return "", errors.Errorf("auths: %v", err)
-	}
-
-	// Get vote details
-	vd, err := p.voteDetails(token)
-	if err != nil {
-		return "", errors.Errorf("voteDetails: %v", err)
-	}
-
-	// Prepare rely
-	dr := ticketvote.DetailsReply{
-		Auths: auths,
-		Vote:  vd,
-	}
-	reply, err := json.Marshal(dr)
-	if err != nil {
-		return "", err
-	}
-
-	return string(reply), nil
-}
-
 // cmdRunoffDetails is an internal plugin command that requests the details of
 // a runoff vote.
 func (p *plugin) cmdRunoffDetails(token []byte) (string, error) {
@@ -795,27 +589,6 @@ func (p *plugin) cmdRunoffDetails(token []byte) (string, error) {
 		Runoff: *srs,
 	}
 	reply, err := json.Marshal(r)
-	if err != nil {
-		return "", err
-	}
-
-	return string(reply), nil
-}
-
-// cmdResults requests the vote objects of all votes that were cast in a ticket
-// vote.
-func (p *plugin) cmdResults(token []byte) (string, error) {
-	// Get vote results
-	votes, err := p.voteResults(token)
-	if err != nil {
-		return "", err
-	}
-
-	// Prepare reply
-	rr := ticketvote.ResultsReply{
-		Votes: votes,
-	}
-	reply, err := json.Marshal(rr)
 	if err != nil {
 		return "", err
 	}
@@ -1012,148 +785,6 @@ func (p *plugin) cmdSubmissions(token []byte) (string, error) {
 	}
 
 	return string(reply), nil
-}
-
-// voteResults returns all votes that were cast in a ticket vote.
-func (p *plugin) voteResults(token []byte) ([]ticketvote.CastVoteDetails, error) {
-	// Retrieve blobs
-	desc := []string{
-		dataDescriptorCastVoteDetails,
-		dataDescriptorVoteCollider,
-	}
-	blobs, err := tstore.BlobsByDataDesc(token, desc)
-	if err != nil {
-		return nil, err
-	}
-
-	// Decode blobs. A cast vote is considered valid only if the vote
-	// collider exists for it. If there are multiple votes using the same
-	// ticket, the valid vote is the one that immediately precedes the vote
-	// collider blob entry.
-	var (
-		// map[ticket]CastVoteDetails
-		votes = make(map[string]ticketvote.CastVoteDetails, len(blobs))
-
-		// map[ticket][]index
-		voteIndexes = make(map[string][]int, len(blobs))
-
-		// map[ticket]index
-		colliderIndexes = make(map[string]int, len(blobs))
-	)
-	for i, v := range blobs {
-		// Decode data hint
-		b, err := base64.StdEncoding.DecodeString(v.DataHint)
-		if err != nil {
-			return nil, err
-		}
-		var dd store.DataDescriptor
-		err = json.Unmarshal(b, &dd)
-		if err != nil {
-			return nil, err
-		}
-		switch dd.Descriptor {
-		case dataDescriptorCastVoteDetails:
-			// Decode cast vote
-			cv, err := convertCastVoteDetailsFromBlobEntry(v)
-			if err != nil {
-				return nil, err
-			}
-
-			// Save index of the cast vote
-			idx, ok := voteIndexes[cv.Ticket]
-			if !ok {
-				idx = make([]int, 0, 32)
-			}
-			idx = append(idx, i)
-			voteIndexes[cv.Ticket] = idx
-
-			// Save the cast vote
-			votes[cv.Ticket] = *cv
-
-		case dataDescriptorVoteCollider:
-			// Decode vote collider
-			vc, err := convertVoteColliderFromBlobEntry(v)
-			if err != nil {
-				return nil, err
-			}
-
-			// Sanity check
-			_, ok := colliderIndexes[vc.Ticket]
-			if ok {
-				return nil, errors.Errorf("duplicate vote "+
-					"colliders found %v", vc.Ticket)
-			}
-
-			// Save the ticket and index for the collider
-			colliderIndexes[vc.Ticket] = i
-
-		default:
-			return nil, errors.Errorf("invalid data descriptor: %v",
-				dd.Descriptor)
-		}
-	}
-
-	for ticket, indexes := range voteIndexes {
-		// Remove any votes that do not have a collider blob
-		colliderIndex, ok := colliderIndexes[ticket]
-		if !ok {
-			// This is not a valid vote
-			delete(votes, ticket)
-			continue
-		}
-
-		// If multiple votes have been cast using the same ticket then
-		// we must manually determine which vote is valid.
-		if len(indexes) == 1 {
-			// Only one cast vote exists for this ticket. This is
-			// good.
-			continue
-		}
-
-		// Sanity check
-		if len(indexes) == 0 {
-			return nil, errors.Errorf("no cast vote index found %v",
-				ticket)
-		}
-
-		log.Tracef("Multiple votes found for a single vote collider %v",
-			ticket)
-
-		// Multiple votes exist for this ticket. The vote that is valid
-		// is the one that immediately precedes the vote collider.
-		// Start at the end of the vote indexes and find the first vote
-		// index that precedes the collider index.
-		var validVoteIndex int
-		for i := len(indexes) - 1; i >= 0; i-- {
-			voteIndex := indexes[i]
-			if voteIndex < colliderIndex {
-				// This is the valid vote
-				validVoteIndex = voteIndex
-				break
-			}
-		}
-
-		// Save the valid vote
-		b := blobs[validVoteIndex]
-		cv, err := convertCastVoteDetailsFromBlobEntry(b)
-		if err != nil {
-			return nil, err
-		}
-		votes[cv.Ticket] = *cv
-	}
-
-	// Put votes into an array
-	cvotes := make([]ticketvote.CastVoteDetails, 0, len(blobs))
-	for _, v := range votes {
-		cvotes = append(cvotes, v)
-	}
-
-	// Sort by ticket hash
-	sort.SliceStable(cvotes, func(i, j int) bool {
-		return cvotes[i].Ticket < cvotes[j].Ticket
-	})
-
-	return cvotes, nil
 }
 
 // voteOptionResults tallies the results of a ticket vote and returns a
@@ -1580,116 +1211,6 @@ func voteIsApproved(vd ticketvote.VoteDetails, results []ticketvote.VoteOptionRe
 	}
 
 	return approved
-}
-
-func convertCastVoteDetailsFromBlobEntry(be store.BlobEntry) (*ticketvote.CastVoteDetails, error) {
-	// Decode and validate data hint
-	b, err := base64.StdEncoding.DecodeString(be.DataHint)
-	if err != nil {
-		return nil, errors.Errorf("decode DataHint: %v", err)
-	}
-	var dd store.DataDescriptor
-	err = json.Unmarshal(b, &dd)
-	if err != nil {
-		return nil, errors.Errorf("unmarshal DataHint: %v", err)
-	}
-	if dd.Descriptor != dataDescriptorCastVoteDetails {
-		return nil, errors.Errorf("unexpected data descriptor: got %v, "+
-			"want %v", dd.Descriptor, dataDescriptorCastVoteDetails)
-	}
-
-	// Decode data
-	b, err = base64.StdEncoding.DecodeString(be.Data)
-	if err != nil {
-		return nil, errors.Errorf("decode Data: %v", err)
-	}
-	digest, err := hex.DecodeString(be.Digest)
-	if err != nil {
-		return nil, errors.Errorf("decode digest: %v", err)
-	}
-	if !bytes.Equal(util.Digest(b), digest) {
-		return nil, errors.Errorf("data is not coherent; got %x, want %x",
-			util.Digest(b), digest)
-	}
-	var cv ticketvote.CastVoteDetails
-	err = json.Unmarshal(b, &cv)
-	if err != nil {
-		return nil, errors.Errorf("unmarshal CastVoteDetails: %v", err)
-	}
-
-	return &cv, nil
-}
-
-func convertVoteColliderFromBlobEntry(be store.BlobEntry) (*voteCollider, error) {
-	// Decode and validate data hint
-	b, err := base64.StdEncoding.DecodeString(be.DataHint)
-	if err != nil {
-		return nil, errors.Errorf("decode DataHint: %v", err)
-	}
-	var dd store.DataDescriptor
-	err = json.Unmarshal(b, &dd)
-	if err != nil {
-		return nil, errors.Errorf("unmarshal DataHint: %v", err)
-	}
-	if dd.Descriptor != dataDescriptorVoteCollider {
-		return nil, errors.Errorf("unexpected data descriptor: got %v, "+
-			"want %v", dd.Descriptor, dataDescriptorVoteCollider)
-	}
-
-	// Decode data
-	b, err = base64.StdEncoding.DecodeString(be.Data)
-	if err != nil {
-		return nil, errors.Errorf("decode Data: %v", err)
-	}
-	digest, err := hex.DecodeString(be.Digest)
-	if err != nil {
-		return nil, errors.Errorf("decode digest: %v", err)
-	}
-	if !bytes.Equal(util.Digest(b), digest) {
-		return nil, errors.Errorf("data is not coherent; got %x, want %x",
-			util.Digest(b), digest)
-	}
-	var vc voteCollider
-	err = json.Unmarshal(b, &vc)
-	if err != nil {
-		return nil, errors.Errorf("unmarshal vote collider: %v", err)
-	}
-
-	return &vc, nil
-}
-
-func convertBlobEntryFromCastVoteDetails(cv ticketvote.CastVoteDetails) (*store.BlobEntry, error) {
-	data, err := json.Marshal(cv)
-	if err != nil {
-		return nil, err
-	}
-	hint, err := json.Marshal(
-		store.DataDescriptor{
-			Type:       store.DataTypeStructure,
-			Descriptor: dataDescriptorCastVoteDetails,
-		})
-	if err != nil {
-		return nil, err
-	}
-	be := store.NewBlobEntry(hint, data)
-	return &be, nil
-}
-
-func convertBlobEntryFromVoteCollider(vc voteCollider) (*store.BlobEntry, error) {
-	data, err := json.Marshal(vc)
-	if err != nil {
-		return nil, err
-	}
-	hint, err := json.Marshal(
-		store.DataDescriptor{
-			Type:       store.DataTypeStructure,
-			Descriptor: dataDescriptorVoteCollider,
-		})
-	if err != nil {
-		return nil, err
-	}
-	be := store.NewBlobEntry(hint, data)
-	return &be, nil
 }
